backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
pydantic_settings==2.4.0
websockets==11.0.3
python-dotenv==1.0.1
openai==1.64.0
